moving pagination types to pagination.type.ts
